Split game logic to hook
